# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Common Login Service 提供用户认证、注册、登录等功能的 API 服务",
        "title": "Common Login Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/auth/google/callback": {
            "get": {
                "description": "Handle Google OAuth callback after user authentication",
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth callback",
                "parameters": [
                    {
                        "type": "string",
                        "description": "OAuth authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Redirect to login page with error"
                    },
                    "400": {
                        "description": "Google login not enabled",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
        "/auth/google/login": {
            "get": {
                "description": "Redirect to Google OAuth login page",
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth login",
                "responses": {
                    "307": {
                        "description": "Redirect to Google OAuth"
                    },
                    "400": {
                        "description": "Google login not enabled",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "Login with Google using ID token from frontend SDK",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Google OAuth API login",
                "parameters": [
                    {
                        "description": "Google login request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.GoogleLoginAPIRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_e54385991_Common-LoginService_internal_service.AuthResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad request or invalid token",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
...
        "/captcha/generate": {
            "post": {
                "description": "Generate a new captcha challenge",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "Generate captcha",
                "responses": {
                    "200": {
                        "description": "Captcha generated",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/internal_handler.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_handler.CaptchaGenerateResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Failed to generate captcha",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        },
...
        "/captcha/verify": {
            "post": {
                "description": "Verify the captcha answer",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "captcha"
                ],
                "summary": "Verify captcha",
                "parameters": [
                    {
                        "description": "Captcha verify request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_handler.CaptchaVerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verification successful",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    },
                    "400": {
                        "description": "Verification failed",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_e54385991_Common-LoginService_internal_model.User": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "display_name": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "last_login_at": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                },
                "vip_level": {
                    "type": "integer"
                }
            }
        },
        "github_com_e54385991_Common-LoginService_internal_service.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/github_com_e54385991_Common-LoginService_internal_model.User"
                }
            }
        },
...
        "internal_handler.CaptchaGenerateResponse": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean",
                    "example": true
                },
                "id": {
                    "type": "string",
                    "example": "captcha-abc123"
                },
                "target": {
                    "type": "string",
                    "example": "请点击图片中的猫"
                }
            }
        },
...
        "internal_handler.CaptchaVerifyRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "example": "captcha-abc123"
                },
                "position": {
                    "type": "integer",
                    "example": 150
                }
            }
        },
...
        "internal_handler.GoogleLoginAPIRequest": {
            "type": "object",
            "required": [
                "id_token"
            ],
            "properties": {
                "id_token": {
                    "type": "string",
                    "example": "google-id-token-xyz"
                }
            }
        },
...
        "internal_handler.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "操作成功"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT token for authorization (format: Bearer {token})",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}